Allow auto-generated classes in repo
This is strictly for enabling CI to be run on this project. The problem
is that for these classes to be generated a DB docker container must be
running, and that was harder than expected to setup using any CI tool.
So this is a emergency solution that hopefully will be replaced in the
future.