Use optionals for indicating request status
Before this change a request could fail and it would throw an unhandled
exception. This change makes sure every exception that can be
thrown when making a request is handled properly. And it's handled using
optionals. An empty optional indicates the request did not return
anything or went bad, which is then handled upstream. Before this commit
the application would eventually fail but this puts it back in business
again.