Use working DB host for running app in IDEA
The commented host is used to reference the docker container name, i.e.
"db". It should be used when starting the containers from the same time
using the docker-compose file (running `make`). But when just running
the application from the IDEA it should reference localhost, which
requires the db to be started manually (`make start-db`).