Throw and handle exception on bad http response
This has been a problem since the beginning and I finally came up with a
good solution. The solution is probably a very default solution for more
experienced developers but for me it was a bit of a revelation.

When the requester throws an exception for getting a bad response,
timing out or getting a 404, the future will complete with an exception.
Prior to this change the future was completed with an empty optional, which does
not complete it. Not sure what happened before really, but it did not
work as intended. So before this change it would get stuck and would not
finish and the consumers would just sit and wait for it to finish.

The exception being thrown is a custom exception, the reasoning behind
that was that I was not sure which one to throw since it can throw
because of multiple reasons, and I did not want to have to handle 4
different exceptions. So I put them all under this exception.