Add PhpInfo plugin and init use of URL class
This commit introduces the PhpInfo plugin, and with that introduction
the Url class is also being used internally. The plan is to use it
everywhere, but for now it's only used internally.

With this change also comes a new type of request called RAW where you
request the raw response object. This is useful if you're interested in
the status code and the content of the response, for example. It's
wrapped in the Url class though, so any data flow needs to be configured
in the Url class.

I did not want to expose the response object immediately since I figured
I could add more useful information to it upon creation. See notes in
Url class.