Let Url class handle all validation and start using it
This commit puts every validation method inside the constructor of the
Url class. So upon creation of the Url instance it validates and
normalises the Url so we know that it's valid and ready to be processed.
No more checking if the Url is valid all over the place, it's now kept
in a single place.

This commit also starts using this class wherever validation was
previously done. Which was one two places. But with this change every
Url flowing through the system is works and looks good!